HealpixMPI.jl provides MPI-parallel versions, through overloads, of most of the other spherical-harmonic related functions of Healpix.jl. Refer to Healpix.jl documentation for their description.

Algebraic operations in harmonic space

HealpixMPI.jl provides overloads of the Base functions +, -, *, /, as well as LinearAlgebra.dot (which embeds and MPI.Allreduce call), allowing to carry out these fundamental operations element-wise in harmonic space directly.

Healpix.almxfl — Function

almxfl(alm::DAlm{S,T,I}, fl::AA) where {S<:Strategy, T<:Number, I<:Integer, AA<:AbstractArray{T,1}}

Multiply a subset of aℓm in the form of DAlm by a vector bℓ representing an ℓ-dependent function, without changing the a_ℓm passed in input.

Arguments

alm::DAlm{S,T,I}: The array representing the spherical harmonics coefficients

fl::AbstractVector{T}: The array giving the factor fℓ by which to multiply aℓm

Returns

Alm{S,T}: The result of aℓm * fℓ.

Healpix.almxfl! — Function

almxfl!(alm::DAlm{S,T,I}, fl::AA) where {S<:Strategy, T<:Number, I<:Integer, AA<:AbstractArray{T,1}}

Multiply IN-PLACE a subset of a_ℓm in the form of DAlm by a vector fl representing an ℓ-dependent function.

Arguments

alms::DAlm{S,T,I}: The subset of spherical harmonics coefficients

fl::AbstractVector{T}: The array giving the factor fℓ by which to multiply aℓm

Base.:+ — Function

+(alm₁::DAlm{S,T,I}, alm₂::DAlm{S,T,I}) where {S<:Strategy, T<:Number, I<:Integer}

Perform the element-wise SUM of two DAlm objects in a_ℓm space. A new DAlm object is returned.

Base.:- — Function

-(alm₁::DAlm{S,T,I}, alm₂::DAlm{S,T,I}) where {S<:Strategy, T<:Number, I<:Integer}

Perform the element-wise SUBTRACTION of two DAlm objects in a_ℓm space. A new DAlm object is returned.

Base.:* — Function

*(alm::DAlm{S,T,I}, fl::AA) where {S<:Strategy, T<:Number, I<:Integer, AA<:AbstractArray{T,1}}
*(fl::AA, alm::DAlm{S,T,I}) where {S<:Strategy, T<:Number, I<:Integer, AA<:AbstractArray{T,1}}

Perform the MULTIPLICATION of a DAlm object by a function of ℓ in a_ℓm space. Note: this consists in a shortcut of almxfl, therefore a new DAlm object is returned.

*(alm₁::DAlm{S,T,I}, alm₂::DAlm{S,T,I}) where {S<:Strategy, T<:Number, I<:Integer}
*(alm₁::DAlm{S,T,I}, c::Number) where {S<:Strategy, T<:Number, I<:Integer}
*(c::Number, alm₁::DAlm{S,T,I}) where {S<:Strategy, T<:Number, I<:Integer}

Perform the element-wise MULTIPLICATION of two DAlm objects or of a DAlm by a constant in a_ℓm space. A new DAlm object is returned.

Base.:/ — Function

/(alm::DAlm{S,T,I}, fl::AA) where {S<:Strategy, T<:Number, I<:Integer, AA<:AbstractArray{T,1}}

Perform an element-wise DIVISION by a function of ℓ in a_ℓm space. Note: this consists in a shortcut of almxfl, therefore a new DAlm object is returned.

/(alm₁::DAlm{S,T,I}, alm₂::DAlm{S,T,I}) where {S<:Strategy, T<:Number, I<:Integer}
/(alm₁::DAlm{S,T,I}, c::Number) where {S<:Strategy, T<:Number, I<:Integer}

Perform the element-wise DIVISION of two DAlm objects or of a DAlm by a constant in a_ℓm space. A new DAlm object is returned.

LinearAlgebra.dot — Function

dot(alm₁::DAlm{S,T,I}, alm₂::DAlm{S,T,I}) where {S<:Strategy, T<:Number, I<:Integer} -> Number

MPI-parallel dot product between two DAlm object of matching size.

Power spectrum

Power spectrum components $C_{\ell}$ are encoded as Vector{T}. HealpixMPI.jl implements overloads of Healpix.jl functions to compute a power spectrum from a set of DAlm (alm2cl) and to generate a set of DAlm from a power spectrum (synalm!).

Healpix.alm2cl — Function

alm2cl(alm₁::DAlm{S,T,I}, alm₂::DAlm{S,T,I}) where {S<:Strategy, T<:Number, I<:Integer} -> Vector{T}
alm2cl(alm::DAlm{S,T,I}) where {S<:Strategy, T<:Number, I<:Integer} -> Vector{T}

Compute the power spectrum $C_{\ell}$ on each MPI task from the spherical harmonic coefficients of one or two fields, distributed as DAlm.

Healpix.synalm! — Function

synalm!(cl::Vector{T}, alm::DAlm{S,N,I}, rng::AbstractRNG) where {S<:Strategy, T<:Real, N<:Number, I<:Integer}
synalm!(cl::Vector{T}, alm::DAlm{S,N,I}) where {S<:Strategy, T<:Real, N<:Number, I<:Integer}

Generate a set of DAlm from a given power spectra array cl. The output is written into the Alm object passed in input. An RNG can be specified, otherwise it's defaulted to Random.GLOBAL_RNG.