Spherical Harmonic Transforms

The main target of HealpixMPI.jl, in terms of run time speed up, are the SHTs, which often represent the "bottle neck" of simulation codes or data analysis pipelines.

As mentioned in the introduction, HealpixMPI.jl relies on ducc's state-of-the-art algorithms for performing the spherical harmonic transforms. In particular, its C++ functions are exploited for the computation of Legandre coefficients from alms and maps and vice versa, while the transposition of such coefficients between MPI tasks is entirely coded in Julia.

The two SHTs are implemented as overloads of Healpix.jl's functions.

Healpix.alm2map! — Function

alm2map!(d_alm::DAlm{S,N,I}, d_map::DMap{S,T,I}, aux_in_leg::StridedArray{Complex{T},3}, aux_out_leg::StridedArray{Complex{T},3}; nthreads::Integer = 0) where {S<:Strategy, N<:Number, T<:Real, I<:Integer}
alm2map!(d_alm::DAlm{S,N,I}, d_map::DMap{S,T,I}; nthreads::Integer = 0) where {S<:Strategy, N<:Number, T<:Real, I<:Integer}

This function performs an MPI-parallel spherical harmonic transform, computing a distributed map from a set of DAlm and places the results in the passed d_map object.

It must be called simultaneously on all the MPI tasks containing the subsets which form exactly the whole map and alm.

It is possible to pass two auxiliary arrays where the Legandre coefficients will be stored during the transform, this avoids allocating extra memory and improves efficiency.

Arguments:

d_alm::DAlm{S,N,I}: the MPI-distributed spherical harmonic coefficients to transform.

d_map::DMap{S,T,I}: the MPI-distributed map that will contain the result.

Optionals:

aux_in_leg::StridedArray{Complex{T},3}: (localnm, totnring, 1) auxiliary matrix for alm-side Legandre coefficients.

aux_out_leg::StridedArray{Complex{T},3}: (totnm, localnring, 1) auxiliary matrix for map-side Legandre coefficients.

Keywords

nthreads::Integer = 0: the number of threads to use for the computation if 0, use as many threads as there are hardware threads available on the system.

Healpix.adjoint_alm2map! — Function

adjoint_alm2map!(d_map::DMap{S,T,I}, d_alm::DAlm{S,N,I}, aux_in_leg::StridedArray{Complex{T},3}, aux_out_leg::StridedArray{Complex{T},3}; nthreads = 0) where {S<:Strategy, N<:Number, T<:Real, I<:Integer}
adjoint_alm2map!(d_map::DMap{S,T,I}, d_alm::DAlm{S,N,I}; nthreads::Integer = 0) where {S<:Strategy, N<:Number, T<:Real, I<:Integer}

This function performs an MPI-parallel spherical harmonic transform Yᵀ on the distributed map and places the results in the passed d_alm object.

It must be called simultaneously on all the MPI tasks containing the subsets which form exactly the whole map and alm.

It is possible to pass two auxiliary arrays where the Legandre coefficients will be stored during the transform, this avoids allocating extra memory and improves efficiency.

Arguments:

d_map::DMap{S,T,I}: the distributed map that must be decomposed in spherical harmonics.

alm::Alm{ComplexF64, Array{ComplexF64, 1}}: the spherical harmonic coefficients to be written to.

Optionals:

aux_in_leg::StridedArray{Complex{T},3}: (localnm, totnring, 1) auxiliary matrix for map-side Legandre coefficients.

aux_out_leg::StridedArray{Complex{T},3}: (totnm, localnring, 1) auxiliary matrix for alm-side Legandre coefficients.

Keywords

nthreads::Integer = 0: the number of threads to use for the computation if 0, use as many threads as there are hardware threads available on the system.

Other Spherical Harmonic functionalities

HealpixMPI.jl provides parallel versions, through overloads, of most of the other spherical-harmonic related functions of Healpix.jl. Refer to Healpix.jl documentation for their description.

Healpix.alm2cl — Function

alm2cl(alm₁::DAlm{S,T,I}, alm₂::DAlm{S,T,I}) where {S<:Strategy, T<:Number, I<:Integer} -> Vector{T}
alm2cl(alm::DAlm{S,T,I}) where {S<:Strategy, T<:Number, I<:Integer} -> Vector{T}

Compute the power spectrum $C_{\ell}$ on each MPI task from the spherical harmonic coefficients of one or two fields, distributed as DAlm.

Healpix.synalm! — Function

synalm!(cl::Vector{T}, alm::DAlm{S,N,I}, rng::AbstractRNG) where {S<:Strategy, T<:Real, N<:Number, I<:Integer}
synalm!(cl::Vector{T}, alm::DAlm{S,N,I}) where {S<:Strategy, T<:Real, N<:Number, I<:Integer}

Generate a set of DAlm from a given power spectra array cl. The output is written into the Alm object passed in input. An RNG can be specified, otherwise it's defaulted to Random.GLOBAL_RNG.

Healpix.almxfl! — Function

almxfl!(alm::DAlm{S,T,I}, fl::AA) where {S<:Strategy, T<:Number, I<:Integer, AA<:AbstractArray{T,1}}

Multiply IN-PLACE a subset of a_ℓm in the form of DAlm by a vector fl representing an ℓ-dependent function.

Arguments

alms::DAlm{S,T,I}: The subset of spherical harmonics coefficients

fl::AbstractVector{T}: The array giving the factor fℓ by which to multiply aℓm