Distributed Classes

As mentioned in the introduction, HealpixMPI has the main purpose of providing an MPI parallelization of the main functionalities of Healpix.jl, distributing maps and harmonic coefficients over the MPI tasks efficiently. This is made possible by the implementation of two data types: DMap and DAlm, mirroring HealpixMap and Alm types of Healpix.jl respectively, and containing a well-defined subset of a map or harmonic coefficients, to be constructed on each MPI task.

HealpixMPI.DMap — Type

struct DMap{T<:Number, I<:Integer}

A subset of a Healpix map, containing only certain rings (as specified in the info field). The type T is used for the value of the pixels in a map, it must be a Number (usually float).

A DMap type contains the following fields:

pixels::Vector{T}: array of pixels composing the subset.

info::GeomInfo: a GeomInfo object describing the HealpixMap subset.

The GeomInfoMPI contained in info must match exactly the characteristic of the Map subset, this is already automatically constructed when MPI.Scatter! is called, reason why this method for initializing a DMap is reccomended.

HealpixMPI.DAlm — Type

struct DAlm{S<:Strategy, T<:Number, I<:Integer}

An MPI-distributed subset of harmonic coefficients a_ℓm, referring only to certain values of m.

The type T is used for the value of each harmonic coefficient, and it must be a Number (one should however only use complex types for this).

A SubAlm type contains the following fields:

alm: the array of harmonic coefficients

info: an AlmInfo object describing the alm subset.

The AlmInfo contained in info must match exactly the characteristic of the Alm subset, this can be constructed through the function make_general_alm_info, for instance.

An instance of DMap (or DAlm) embeds, whithin the field info, a GeomInfoMPI (or AlmInfoMPI) object. These latter, in turn, contain all the necessairy information about:

The whole map geometry (or the whole set of harmonic coefficients).

The composition of the local subset.

the MPI communicator.

HealpixMPI.GeomInfoMPI — Type

struct GeomInfoMPI{I <: Integer, T <: Real}

Information describing an MPI-distributed subset of a HealpixMap, contained in a DMap.

A GeomInfoMPI type contains:

comm: MPI communicator used.

nside: NSIDE parameter of the whole map.

maxnr: maximum number of rings in the subsets, over the tasks involved.

thetatot: array of the colatitudes of the whole map ordered by task first and RR within each task

rings: array of the ring indexes (w.r.t. the whole map) contained in the subset.

rstart: array containing the 1-based index of the first pixel of each ring contained in the subset.

nphi: array containing the number of pixels in every ring contained in the subset.

theta: array of colatitudes (in radians) of the rings contained in the subset.

phi0: array containing the values of the azimuth (in radians) of the first pixel in every ring.

HealpixMPI.AlmInfoMPI — Type

struct AlmInfoMPI{I <: Integer}

Information describing an MPI-distributed subset of Alm, contained in a DAlm.

An AlmInfoMPI type contains:

comm: MPI communicator used.

lmax: the maximum value of $ℓ$.

mmax: the maximum value of $m$ on the whole Alm set.

maxnm: maximum value (over tasks) of nm (number of m values).

mval: array of values of $m$ contained in the subset.

mstart: array hypothetical indexes of the harmonic coefficient with ℓ=0, m. #FIXME: for now 0-based

Initializing a Distributed type

The recommended way to construct a local subset of a map or harmonic coefficients, is to start with an instance of HealpixMap (in RingOrder) or Alm on the root task, and call one of the apposite overloads of the standard MPI.Scatter! function, provided by HealpixMPI.jl. Such function would in fact save the user the job of constructing all the required ancillary information describing the data subset, doing so through efficient and tested methods.

MPI.Scatter! — Function

Scatter!(in_alm::Alm{T,Array{T,1}}, out_d_alm::DAlm{T}, comm::MPI.Comm; root::Integer = 0, clear::Bool = false) where {S<:Strategy, T<:Number, I<:Integer}
Scatter!(in_alm::Nothing, out_d_alm::DAlm{T}, comm::MPI.Comm; root::Integer = 0, clear::Bool = false) where {S<:Strategy, T<:Number, I<:Integer}
Scatter!(in_alm, out_d_alm::DAlm{S,T,I}, comm::MPI.Comm; root::Integer = 0, clear::Bool = false) where {S<:Strategy, T<:Number, I<:Integer}

Distributes the Alm object passed in input on the root task overwriting the DAlm objects passed on each task, according to the specified strategy (by default ":RR" for Round Robin).

As in the standard MPI function, the in_alm in input can be nothing on non-root tasks, since it will be ignored anyway.

If the keyword clear is set to true it frees the memory of each task from the (potentially bulky) Alm object.

Arguments:

in_alm::Alm{T,Array{T,1}}: Alm object to distribute over the MPI tasks.

out_d_alm::DAlm{T}: output DAlm object.

Keywords:

root::Integer: rank of the task to be considered as "root", it is 0 by default.

clear::Bool: if true deletes the input Alm after having performed the "scattering".

Scatter!(in_map::HealpixMap{T,RingOrder,Array{T,1}}, out_d_map::DMap{S,T,I}, strategy::Symbol, comm::MPI.Comm; root::Integer = 0, clear::Bool = false) where {T <: Number, I <: Integer}
Scatter!(nothing, out_d_map::DMap{S,T,I}, strategy::Symbol, comm::MPI.Comm; root::Integer = 0, clear::Bool = false) where {T <: Number, I <: Integer}

Distributes the HealpixMap object passed in input on the root task overwriting the DMap objects passed on each task, according to the specified strategy (by default ":RR" for Round Robin).

As in the standard MPI function, the in_map in input can be nothing on non-root tasks, since it will be ignored anyway.

If the keyword clear is set to true it frees the memory of each task from the (potentially bulky) HealpixMap object.

Arguments:

in_map::HealpixMap{T,RingOrder,Array{T,1}}: HealpixMap object to distribute over the MPI tasks.

out_d_map::DMap{S,T,I}: output DMap object.

Keywords:

root::Integer: rank of the task to be considered as "root", it is 0 by default.

clear::Bool: if true deletes the input map after having performed the "scattering".

While distributing a set of harmonic coefficients means that each MPI task will host a DAlm object containing only the coefficients corresponding to some specific values of m, the distribution of a map is performed by rings. Each MPI task will then host a DMap object containing only the pixels composing some specified rings of the entire HealpixMap. Note that, for spherical harmonic transforms efficiency, it is recommended to assign pairs of rings with same latitude (i.e. symmetric w.r.t. the equator) to the same task, in order to preserve the geometric symmetry of the map.

The following example shows the standard way to initialize a DAlm object through a round robin strategy (see the paragraph Distributing Strategy for more details about this).

using HealpixMPI

MPI.Init()
comm = MPI.COMM_WORLD

alm = Alm(5, 5, randn(ComplexF64, numberOfAlms(5))) #inizialize random Healpix Alm
d_alm = DAlm{RR}() #inizialize empty DAlm to be filled according to RR strategy

MPI.Scatter!(alm, d_alm, comm) #fill d_alm

Gathering data

Analogously to MPI.Scatter!, HealpixMPI.jl also provides overloads of MPI.Gather! (and MPI.Allgather!). These latter allow to re-group subsets of map or alm into a HealpixMap or Alm only on the root task (or on every MPI task involved).

MPI.Gather! — Function

Gather!(in_d_alm::DAlm{S,T,I}, out_alm::Alm{T,Array{T,1}}, comm::MPI.Comm; root::Integer = 0, clear::Bool = false) where {S<:Strategy, T<:Number, I<:Integer}
Gather!(in_d_alm::DAlm{S,T,I}, out_alm::Nothing, comm::MPI.Comm; root::Integer = 0, clear::Bool = false) where {S<:Strategy, T<:Number, I<:Integer}

Gathers the DAlm objects passed on each task overwriting the Alm object passed in input on the root task according to the specified strategy (by default :RR for Round Robin). Note that the strategy must match the one used to "scatter" the a_lm.

As in the standard MPI function, the out_alm can be nothing on non-root tasks, since it will be ignored anyway.

If the keyword clear is set to true it frees the memory of each task from the (potentially bulky) DAlm object.

Arguments:

in_d_alm::DAlm{T}: DAlm object to gather from the MPI tasks.

out_d_alm::Alm{T,Array{T,1}}: output Alm object.

Keywords:

strategy::Symbol: Strategy to be used, by default :RR for "Round Robin".

root::Integer: rank of the task to be considered as "root", it is 0 by default.

clear::Bool: if true deletes the input Alm after having performed the "scattering".

Gather!(in_d_map::DMap{T, I}, out_map::HealpixMap{T,RingOrder,Array{T,1}}, strategy::Symbol, comm::MPI.Comm; root::Integer = 0, clear::Bool = false)
Gather!(in_d_map::DMap{T, I}, out_map::Nothing, strategy::Symbol, comm::MPI.Comm; root::Integer = 0, clear::Bool = false)

Gathers the DMap objects passed on each task overwriting the HealpixMap object passed in input on the root task according to the specified strategy (by default :RR for Round Robin). Note that the strategy must match the one used to "scatter" the map.

As in the standard MPI function, the out_map can be nothing on non-root tasks, since it will be ignored anyway.

If the keyword clear is set to true it frees the memory of each task from the (potentially bulky) DMap object.

Arguments:

in_d_map::DMap{T, I}: DMap object to gather from the MPI tasks.

out_map::HealpixMap{T,RingOrder,Array{T,1}}: output Map object.

Keywords:

strategy::Symbol: Strategy to be used, by default :RR for "Round Robin".

root::Integer: rank of the task to be considered as "root", it is 0 by default.

clear::Bool: if true deletes the input DMap after having performed the "scattering".

MPI.Allgather! — Function

Allgather!(in_d_alm::DAlm{S,T,I}, out_alm::Alm{T,Array{T,1}}; clear::Bool = false) where {S<:Strategy, T<:Number, I<:Integer}

Gathers the DAlm objects passed on each task overwriting the Alm object passed in input on the root task according to the specified strategy (by default :RR for Round Robin). Note that the strategy must match the one used to "scatter" the a_lm.

As in the standard MPI function, the out_alm can be nothing on non-root tasks, since it will be ignored anyway.

If the keyword clear is set to true it frees the memory of each task from the (potentially bulky) DAlm object.

Arguments:

in_d_alm::DAlm{T}: DAlm object to gather from the MPI tasks.

out_d_alm::Alm{T,Array{T,1}}: output Alm object.

Keywords:

strategy::Symbol: Strategy to be used, by default :RR for "Round Robin".

clear::Bool: if true deletes the input Alm after having performed the "scattering".

Allgather!(in_d_map::DMap{S,T,I}, out_map::HealpixMap{T,RingOrder,Array{T,1}}, strategy::Symbol, comm::MPI.Comm; clear::Bool = false) where {T <: Number}

Gathers the DMap objects passed on each task overwriting the out_map object passed in input on EVERY task according to the specified strategy (by default :RR for Round Robin). Note that the strategy must match the one used to "scatter" the map.

If the keyword clear is set to true it frees the memory of each task from the (potentially bulky) DMap object.

Arguments:

in_d_map::DMap{S,T,I}: DMap object to gather from the MPI tasks.

out_d_map::HealpixMap{T,RingOrder,Array{T,1}}: output HealpixMap object to overwrite.

Keywords:

clear::Bool: if true deletes the input Alm after having performed the "scattering".

Distributing Strategy

It is also worth mentioning that one could find many different strategies to distribute a set of data over multiple MPI tasks. So far, the only one implemented in HealpixMPI.jl, which should guarantee an adequate work balance between tasks, is the so-called "round robin" strategy: assuming $N$ MPI tasks, the map is distributed such that task $i$ hosts the map rings $i$, $i + N$, $i + 2N$, etc. (and their counterparts on the other hemisphere). Similarly, for the spherical harmonic coefficients, task $i$ would hold all coefficients for $m = i$, $i + N$, $i + 2 N$, etc.

The strategy is intrinsically specified in a DMap or DAlm instance through an abstract type (e.g. RR), inherited from a super-type Strategy; in the same way as the pixel ordering is specified in a HealpixMap in Healpix.jl.

This kind of solution allows for two great features:

An efficient and fast multiple-dispatch, allowing a function to recognize the distribution strategy used on data structure without the usage of any if statement.

Allows to add other distributing strategies if needed for future developments by simply adding an inherited type in the source file strategy.jl with a single line of code:

abstract type NewStrat<:Strategy end